--- a/backend/ee/onyx/server/tenants/router.py
+++ b/backend/ee/onyx/server/tenants/router.py
@@ -6,10 +6,9 @@ from fastapi import HTTPException
 from pydantic import BaseModel
 
 from ee.onyx.server.tenants.provisioning import complete_tenant_setup
-from ee.onyx.server.tenants.user_mapping import get_tenant_id_for_email
-from onyx.auth.users import current_user
-from onyx.auth.users import exceptions
-from onyx.db.models import User
+from onyx.auth.users import optional_minimal_user
+from onyx.db.models import MinimalUser
+from shared_configs.contextvars import get_current_tenant_id
 
 logger = logging.getLogger(__name__)
 
@@ -23,25 +22,18 @@ class CompleteTenantSetupRequest(BaseModel):
 @router.post("/complete-setup")
 async def api_complete_tenant_setup(
     request: CompleteTenantSetupRequest,
-    user: User = Depends(current_user),
-) -> dict:
+    user: MinimalUser = Depends(optional_minimal_user),
+) -> None:
     """Complete the tenant setup process for a user.
 
     This endpoint is called from the frontend after user creation to complete
     the tenant setup process (migrations, seeding, etc.).
     """
-    if not user.is_admin and user.email != request.email:
-        raise HTTPException(
-            status_code=403, detail="You can only complete setup for your own tenant"
-        )
+
+    tenant_id = get_current_tenant_id()
 
     try:
-        tenant_id = get_tenant_id_for_email(request.email)
-    except exceptions.UserNotExists:
-        raise HTTPException(status_code=404, detail="User or tenant not found")
-
-    try:
-        await complete_tenant_setup(tenant_id, request.email)
+        await complete_tenant_setup(tenant_id)
         return {"status": "success"}
     except Exception as e:
         logger.error(f"Failed to complete tenant setup: {e}")