… web/src/app/auth/signup/page.tsx
… web/src/app/auth/waiting-on-setup/WaitingOnSetup.tsx
… web/src/app/auth/waiting-on-setup/page.tsx
… web/src/app/layout.tsx
… web/src/lib/tenant.ts
… web/src/lib/userSS.ts