## Description

Get user to a loading screen as quickly as possible. Provision until we have partial tables and then finish the auth.

| Approach | Full initial provisioning process | Alembic migrations |
|----------|-----------------------------------|-------------------|
| This approach | 0.22 seconds | 0.21 seconds |
| Existing approach | 9.1 seconds | 1.68 seconds |

## How Has This Been Tested?

[Describe the tests you ran to verify your changes]

## Backporting (check the box to trigger backport action)

Note: You have to check that the action passes, otherwise resolve the conflicts manually and tag the patches.

- [ ] This PR should be backported (make sure to check that the backport attempt succeeds)
- [ ] [Optional] Override Linear Check